refactor (step2) : CardDeck으로 CARD_SET 이동
