+++ src/main/kotlin/blackjack/domain/CardDeck.kt
@@ -1,6 +1,5 @@
 package blackjack.domain
 
-import blackjack.domain.Card.Companion.CARD_SET
 import java.util.LinkedList
 import java.util.Queue
 
@@ -16,5 +15,20 @@ data class CardDeck(private val cardQueue: Queue<Card> = LinkedList(CARD_SET.shu
     private fun resetQueue() {
         cardQueue.clear()
         cardQueue.addAll(CARD_SET.shuffled())
+    }
+
+    companion object {
+        private val ALL_CARDS = NumberType
+            .values()
+            .flatMap(::generateSuitNumberPair)
+            .associateWith { Card(it.first, it.second) }
+
+        private fun generateSuitNumberPair(numberType: NumberType): List<Pair<SuitType, NumberType>> = SuitType
+            .values()
+            .map { suitType ->
+                suitType to numberType
+            }
+
+        val CARD_SET = ALL_CARDS.values.toSet()
     }
 }
